+++ client/src/components/Sidebar/Sidebar.js
@@ -58,7 +58,7 @@ class Sidebar extends Component {
           </a>
         </div>
         <a href="/" className="brand-logo center">
-        {<img src={ require('../public/logo.png')} width = "50px" height="50px" />}
+        {<img src={ require('../../public/logo.png')} width = "50px" height="50px" />}
           Split It
         </a>
       </div>
